Move `util.ts` into `shared`
