+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/data-type.ts
@@ -5,10 +5,10 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../../util";
 import { DataTypeWithMetadata } from "../../../types/ontology/data-type.js";
 import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
 import { isDataTypeVertex } from "../../../types/subgraph/vertices.js";
+import { typedValues } from "../../../util";
 
 /**
  * Returns all `DataTypeWithMetadata`s within the vertices of the subgraph
