+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/entity-type.ts
@@ -5,10 +5,10 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../../util";
 import { EntityTypeWithMetadata } from "../../../types/ontology/entity-type.js";
 import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
 import { isEntityTypeVertex } from "../../../types/subgraph/vertices.js";
+import { typedValues } from "../../../util";
 
 /**
  * Returns all `EntityTypeWithMetadata`s within the vertices of the subgraph
