+++ libs/@blockprotocol/graph/src/shared/stdlib/subgraph/element/property-type.ts
@@ -5,10 +5,10 @@ import {
   VersionedUri,
 } from "@blockprotocol/type-system/slim";
 
-import { typedValues } from "../../../../util";
 import { PropertyTypeWithMetadata } from "../../../types/ontology/property-type.js";
 import { OntologyTypeVertexId, Subgraph } from "../../../types/subgraph.js";
 import { isPropertyTypeVertex } from "../../../types/subgraph/vertices.js";
+import { typedValues } from "../../../util";
 
 /**
  * Returns all `PropertyTypeWithMetadata`s within the vertices of the subgraph
