+++ libs/@blockprotocol/graph/src/shared/util.ts
